Crate janus_collector

A DAP-PPM collector

This library implements the collector role of the DAP-PPM protocol. It works in concert with two DAP-PPM aggregator servers to compute a statistical aggregate over data from many clients, while preserving the privacy of each client’s data.

Examples

use janus_collector::{AuthenticationToken, Collector, CollectorParameters, default_http_client};
use janus_core::{hpke::generate_hpke_config_and_private_key};
use janus_messages::{
    Duration, HpkeAeadId, HpkeConfig, HpkeConfigId, HpkeKdfId, HpkeKemId, Interval, TaskId,
    Time, Query,
};
use prio::vdaf::prio3::Prio3;
use rand::random;
use url::Url;

// Supply DAP task paramenters.
let task_id = random();
let hpke_keypair = janus_core::hpke::generate_hpke_config_and_private_key(
    HpkeConfigId::from(0),
    HpkeKemId::X25519HkdfSha256,
    HpkeKdfId::HkdfSha256,
    HpkeAeadId::Aes128Gcm,
);
let parameters = CollectorParameters::new(
    task_id,
    "https://example.com/dap/".parse().unwrap(),
    AuthenticationToken::Bearer(b"my-authentication-token".to_vec()),
    hpke_keypair.config().clone(),
    hpke_keypair.private_key().clone(),
);

// Supply a VDAF implementation, corresponding to this task.
let vdaf = Prio3::new_count(2).unwrap();
// Use the default HTTP client as-is.
let http_client = default_http_client().unwrap();
let collector = Collector::new(parameters, vdaf, http_client);

// Specify the time interval over which the aggregation should be calculated.
let interval = Interval::new(
    Time::from_seconds_since_epoch(1_656_000_000),
    Duration::from_seconds(3600),
)
.unwrap();
// Make the requests and retrieve the aggregated statistic.
let aggregation_result = collector.collect(Query::new_time_interval(interval), &()).await.unwrap();

Structs

Collection

The result of a collection operation.

Collector

A DAP collector.

CollectorParameters

The DAP collector’s view of task parameters.

Enums

AuthenticationToken

Different modes of authentication supported by Janus for either sending requests (e.g., leader to helper) or receiving them (e.g., collector to leader).

Error

Errors that may occur when performing collections.

Functions

default_http_client

Construct a reqwest::Client suitable for use in a DAP Collector.